@@ -1,20 +1,47 @@
+import React, { useState } from "react";
+import Header from "./components/Header/Header";
 import About from "./components/About/About";
 import Contact from "./components/Contact/Contact";
-import Home from "./components/Home/Home";
-import Navbar from "./components/Navbar/Navbar";
-import Skills from "./components/Skills/Skills";
-import Works from "./components/Works/Works";
+import Portfolio from "./components/Projects/Project";
+import Resume from "./components/Resume/Resume";
+import Footer from "./components/Footer/Footer";
+import { Helmet } from "react-helmet";
+import "./App.css";
 
-function App() { 
+function App() {
+  const [currentTab, handleTabChange] = useState("about");
+
+  // This method is checking to see what the value of `currentTab` is. Depending on the value of currentPage, we return the corresponding component to render.
+  const renderTab = () => {
+    if (currentTab === "About") {
+      return <About />;
+    }
+    if (currentTab === "Contact") {
+      return <Contact />;
+    }
+    if (currentTab === "Portfolio") {
+      return <Portfolio />;
+    }
+    if (currentTab === "Resume") {
+      return <Resume />;
+    }
+    return <About />;
+  };
+
   return (
-    <div>
-      <Navbar />
-      <Home />
-      <About />
-      <Skills />
-      <Works />
-      <Contact />
-    </div>
+    <>
+      <Helmet>
+        <title>Mohamed's Portfolio | {currentTab} </title>
+      </Helmet>
+      ;
+      <Header
+        currentTab={currentTab}
+        handleTabChange={handleTabChange}
+      ></Header>
+      <main>{renderTab()}</main>
+      <Footer></Footer>
+    </>
   );
 }
+
 export default App;
