@@ -1,42 +1,24 @@
 import React from "react";
-
-const About = () => { 
+// import profileImage from "../../assets/banner.png";
+function About() {
   return (
-    <div
-      name="about"
-      id="about"
-      className="w-full h-screen bg-[#0a192f] text-gray-300"
-    >
-      <div className="flex flex-col justify-center items-center w-full h-full">
-        <div className=" py-16 rounded-md bg-cyan-800 flex flex-col justify-center items-center w-4/6">
-          <div className="max-w-[1000px] w-full grid grid-cols-2 gap-8 mb-4">
-            <div className="sm:text-right pb-8 pl-4">
-              <p className="text-4xl font-bold inline border-b-4 border-cyan-500">
-                About
-              </p>
-            </div>
-            <div></div>
-          </div>
-          <div className="max-w-[1000px] w-full grid sm:grid-cols-2 gap-8 px-4">
-            <div className="sm:text-right text-4xl font-bold">
-              <p>
-                Hi. I'm Taminoturoko Briggs, nice to meet you. Please take a
-                look around.
-              </p>
-            </div>
-            <div>
-              <p>
-                {" "}
-                A software developer with experience in building Responsive and
-                Scalable Web apps. I am well-knowledged in UI/UX principles and
-                practices. In addition to software development, I am also a
-                technical writer--simplifying topics/concepts on the web.
-              </p>
-            </div>
-          </div>
+    <section id="about-me" className="my-5 intro">
+      <div class="flex-row">
+        <h1 id="about" className="section-title primary-border">
+          About Me
+        </h1>
+      </div>
+      <div className="flex-row">
+        <div class="intro-info">
+          <p>
+          Greetings, my name is Mohamed Abdullahi and I am a resident of Ottawa, Canada. 
+          I hold a Bachelor of Science degree in Information Technology from the University of Mount Kenya, which I completed in 2015. Additionally, I have completed the Willis Cyber Security Analyst program in 2021, 
+          and I have also graduated from the Full Stack Developer Bootcamp at Carleton University- Ottawa where I gained expertise in building both frontend and backend web Applications.
+          </p>
         </div>
       </div>
-    </div>
+    </section>
   );
-};
+}
+
 export default About;
